Many changes, only a few of which are mentioned here (yes, I should have used feature branches):   Add a bunch more unit tests.   Initial work on rendering text along a line.   Initial work on supporting SDF glyphs.   Add RenderTarget support for pushing/popping transforms.   Regenerate vector_tile.pb.h/cc and glyphs.pb.h/cc using latest version   of protoc.   Use std::vector<uint32_t> for Bitmap rather than   std::vector<std::byte> - uisng std::byte is more flexible but it could   potentially cause alignment problems.

+++ core/Bitmap.cpp
@@ -24,7 +24,7 @@ namespace core::bitmap
 		{
 			int size = x*y*4;
 
-			std::span<const std::byte> span { reinterpret_cast<const std::byte*>(bitmapData), static_cast<size_t>(size) };
+			std::span<const std::uint32_t> span { reinterpret_cast<const std::uint32_t*>(bitmapData), static_cast<size_t>(size) };
 			Bitmap bitmap(x, y, Format::RGBA, span);
 
 			stbi_image_free(bitmapData);
